requirements.txt
Django==3.2
django-extensions==3.1.2
djangorestframework==3.12.4
django-crispy-forms==1.11.2
django-import-export==2.5.0
tablib==3.0.0